Modificaciones Oficina

+++ SpringBoot/DemoBoot01/src/main/java/com/example/spring/boot/service/BankService.java
@@ -6,7 +6,7 @@ import com.example.spring.boot.model.Bank;
 
 
 public interface BankService {
-	Bank findById(long id);
+	Bank findById(Integer id);
 	Bank findByName(String name);
 	List<Bank> findAllBanks();
 }
